rename remaining halochest to halohalo
